% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segregating_sites.R
\name{discph}
\alias{discph}
\title{Computes P and T* matrix, initial probabilities and defect for some frequency count (singleton, doubleton, ...)}
\usage{
iton_mats(n)

for singletons
}
\description{
Computes P and T* matrix, initial probabilities and defect for some frequency count (singleton, doubleton, ...)
}
