% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_ph.R
\name{ddphtype}
\alias{ddphtype}
\alias{pdphtype}
\title{Density, distribution function, quantile function and random generation for
the discrete phase-type distribution}
\usage{
ddphtype(x, subint_mat, init_probs)

pdphtype(x, subint_mat, init_probs)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{n_samples}{number of samples to draw.}

\item{n_OTU}{number of OTUs to consider}

\item{granularity}{distance between numbers drawable}

\item{type}{type of calculation ("T_MRCA" and later S_total)}
}
\description{
Density, distribution function, quantile function and random generation for
the discrete phase-type distribution
}
\section{Functions}{
\itemize{
\item \code{ddphtype}: Density function.

\item \code{pdphtype}: Distribution function.
}}

