% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dphase_type.R
\name{dphase_type}
\alias{dphase_type}
\alias{mean.dphase_type}
\alias{var.dphase_type}
\alias{summary.dphase_type}
\title{\code{dphase_type} class}
\usage{
dphase_type(n, init_probs = NULL, itons = NULL, theta = 2, moment = F, tail_stat = F)

## S3 method for class 'dphase_type'
mean(object)

## S3 method for class 'dphase_type'
var(object)

## S3 method for class 'dphase_type'
summary(object)
}
\arguments{
\item{n}{integer larger than 1.}

\item{init_probs}{vector, a one-row matrix or \code{NULL} (default).}

\item{itons}{integer between 1 and n-1, or \code{NULL} (default).}

\item{theta}{numeric.}

\item{object}{an object of class \code{dphase_type}.}

\item{type}{\code{'T_MRCA'}, \code{'T_Total'} or \code{NULL} (default).}
}
\description{
Description of the class \code{phase_type}, which represents discrete phase-type
distributions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{mean}: mean of the discrete phase-type distribution.

\item \code{var}: variance of the discrete phase-type distribution.

\item \code{summary}: summary of the discrete phase-type distribution.
}}

