% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generate_subint_mat}
\alias{generate_subint_mat}
\title{Helper functions to generate inputs for the phasetype-related functions.}
\usage{
rphtype(n, init_probs, subint_mat)
}
\arguments{
\item{n}{number of samples.}

\item{type}{string denoting the type of PH-related subintensity matrix wanted: T_MRCA, S_total}
}
\description{
Helper functions to generate inputs for the phasetype-related functions.
}
\section{Functions}{
\itemize{
\item \code{generate_subint_mat}: Generates a subintensity matrix from a specified number of samples.
Currently only supports generating the subintensity matrix used in calculation
of the time to the most recente common ancestor (MRCA)
}}

