% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_ph_functions.R
\name{iton_mats}
\alias{iton_mats}
\title{iton_mats}
\usage{
iton_mats(n)

for singletons
}
\description{
Computes T* matrix, initial probabilities and defect for some frequency count (singleton, doubleton, ...)
}
