% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_phase_type_separated.R
\name{kingsman}
\alias{kingsman}
\title{Simulate Matrices from Kingsman Coalescent}
\usage{
kingsman(n)
}
\arguments{
\item{n}{sample size (should be a positive integer)}
}
\value{
A `mult_phase_type` object containing the subintensity matrix, reward matrix and vector of initial probabilities
}
\description{
Wrapper function that computes the matrices descrbining Kingsman coalescent ready for reward transformation
}
\examples{
kingsman(4)

}
