% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_type_constructor.R
\name{phase_type}
\alias{phase_type}
\alias{summary.phase_type}
\alias{mean.phase_type}
\alias{var.phase_type}
\title{\code{phase_type} class}
\usage{
phase_type(type = NULL, n = NULL,
           subint_mat = NULL, init_probs = NULL)

## S3 method for class 'phase_type'
summary(object)

## S3 method for class 'phase_type'
mean(object)

## S3 method for class 'phase_type'
var(object)
}
\arguments{
\item{type}{\code{'T_MRCA'}, \code{'T_Total'} or \code{NULL} (default).}

\item{n}{integer larger than 1 or \code{NULL} (default).}

\item{subint_mat}{matrix or \code{NULL} (default).}

\item{init_probs}{vector, a one-row matrix or \code{NULL} (default).}

\item{object}{an object of class \code{phase_type}.}
}
\description{
Description of the class \code{phase_type}, which represents continuous phase-type
distributions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: summary of the continuous phase-type distribution.

\item \code{mean}: mean of the continuous phase-type distribution.

\item \code{var}: variance of the continuous phase-type distribution.
}}

