% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segregating_sites.R
\name{rewardtransformparm}
\alias{rewardtransformparm}
\title{RewardTransformParm}
\usage{
rewardtransformparm(n, init_probs, subint_mat)
}
\description{
Function to compute reward transformed matrix T* as well as vector of initial probabilities and the defect
}
