% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RewTransform.R
\name{RewTransform}
\alias{RewTransform}
\title{Discretize MPH object with a reward vector}
\usage{
RewTransform(mph_obj, rewards, theta)
}
\arguments{
\item{mph_obj}{multivariate phase-type object generated either from mult_phase_type or kingsman function}

\item{rewards}{vector of non negative numbers}

\item{theta}{mutation parameter (positive number)}
}
\value{
A list containing keys `subint_mat`, `init_probs` and `defect`
}
\description{
Reward transformation of continuous phase-type distribution into a discrete phase-type distribution.
}
\examples{
n = 5
mphobj = kingsman(n)
reward_length = length(mphobj$init_probs)
transformed = RewTransform(mphobj, 0:(reward_length-1), 2)
transformed$subint_mat

}
