% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.r
\name{sanity_check_rphdph}
\alias{sanity_check_rphdph}
\title{Sanity checks, showing that everything works as intended}
\usage{
sanity_check_rphdph = function(OTUs, draws, granularity)
}
\arguments{
\item{OTUs}{number of operational taxonomical units}

\item{draws}{number of random draws from rphtype wanted}

\item{granularity}{the difference between the numbers put into dphtype()}

\item{test}{unknown}
}
\description{
Sanity checks, showing that everything works as intended
}
\section{Functions}{
\itemize{
\item \code{sanity_check_rphdph}: Makes a simple comparison between theoretical and randomly sampled data from dphtype() and rphtype, respectively
}}

