% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\name{sfs}
\alias{sfs}
\title{Site Frequency Spectrum}
\usage{
sfs(n, theta = 2, init_probs = NULL, plot = FALSE)
}
\arguments{
\item{n}{sample size}

\item{theta}{population mutation parameter}

\item{init_probs}{vector of initial probabilities. There is no need to supply entire vector if most of its elements are zero. The function computes the rest automatically}

\item{plot}{logical}
}
\description{
Returns a list of expected counts of itons up to n-1 or a plot if specified. Works with different values of `theta` (population mutation parameter)
as well as with custom vector of initial probabilities.
}
\examples{
sfs(n = 5)
sfs(5, theta = 3)
sfs(10, 2, init_probs = c(0.8, 0.2))

}
