% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mrca.R
\name{t_mrca}
\alias{t_mrca}
\title{Time until the most recent common ancestor}
\usage{
t_mrca(n)
}
\arguments{
\item{n}{positive integer larger than 1.}
}
\description{
The function creates a continuous phase-type representation of the time until the most recent
common ancestor for a specified number of sequences and following Kingsman's coalescent model.
}
\examples{
ph_mrca <- t_mrca(4)
summary(ph_mrca)

}
