% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_phase_type_separated.R
\name{tailstat}
\alias{tailstat}
\title{Tail Statistic}
\usage{
tail_stat(mph_obj, k, theta)
}
\arguments{
\item{mph_obj}{Multivariate Phase Type object generated either from mult_phase_type or kingsman function}

\item{k}{Minimum Frequency count (positive integer)}

\item{theta}{mutation parameter (positive)}
}
\value{
A `disc_phase_type` object containing subintensity matrix (P), vector of initial probabilities (alpha) and defect (probability of not entering any transient
state prior to absorption)
}
\description{
Generate subintensity matrix, vector of initial probabilites and defect for combination of frequency counts higher than "k" (tail statistic)
}
