% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_type.R
\name{var}
\alias{var}
\title{Variance of phase_type distributions}
\usage{
var(obj)
}
\arguments{
\item{obj}{a phase_type or disc_phase_type object.}
}
\description{
It calculates the variance of continuous and discrete phase-type distributions,
represented by the \code{phase_type} and the \code{disc_phase_type} classes
respectively
}
