% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_phase_type.R
\name{disc_phase_type}
\alias{disc_phase_type}
\title{\code{disc_phase_type} class}
\usage{
disc_phase_type(n = NULL, itons = NULL,
            theta = 2, tail_stat = F,
            subint_mat = NULL, init_probs = NULL)
}
\arguments{
\item{n}{integer larger than 1, or \code{NULL} (default).}

\item{itons}{integer between 1 and n-1, or \code{NULL} (default).}

\item{theta}{numeric.}

\item{tail_stat}{logical.}

\item{subint_mat}{matrix or \code{NULL} (default).}

\item{init_probs}{vector, a one-row matrix or \code{NULL} (default).}
}
\description{
Description of the class \code{disc_phase_type}, which represents discrete phase-type
distributions.
}
\details{
\code{disc_phase_type} is the generator function for the discrete phase-type distribution class
of the same name, which inherits from \code{list}. This function can easily produce phase-type
representations of common statistics in genomics, such as the total number of segregating sites,
the number of i-tons (singletons, doubletons, etc.) and the tail statistic.

The discrete phase-type distribution can be generated in two different ways:
\itemize{
  \item By specifying whether the user wants a  discrete phase-type representation
  of the total number of segregating sites (default), the number of i-tons, or the tail
  statistic for a certain number of sequences \code{n}.
  \item By supplying a user-defined sub-intensity matrix, with optional initial
  probabilities.
}

See examples for further explanation on the usage.
}
\examples{
# Total number of segregating sites
disc_phase_type(4)
disc_phase_type(5, theta=0.5)

# Below examples are not implemented:
# Number of singletons
disc_ph_example <- disc_phase_type(4, itons=1)
mean(disc_ph_example)
var(disc_ph_example)
summary(disc_ph_example)

# Tail statistic
disc_ph_tail <- disc_phase_type(4, itons=2, tail_stat=TRUE)
summary(disc_ph_tail)

}
