% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dphtype.R
\name{dphtype}
\alias{dphtype}
\alias{qphtype}
\alias{pphtype}
\alias{rphtype}
\title{Density, distribution function, quantile function and random generation for
the continuous phase-type distribution}
\usage{
dphtype(x, init_probs, subint_mat)

qphtype(q, init_probs, subint_mat)

pphtype(p, init_probs, subint_mat)

rphtype(n_samples, n_OTU, granularity = 0.01, type = "T_MRCA")
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{init_probs}{vector of initial probabilities.}

\item{subint_mat}{subintensity matrix.}

\item{p}{vector of probabilities.}

\item{n_samples}{number of samples to draw.}

\item{n_OTU}{number of OTUs to consider}

\item{granularity}{distance between numbers drawable}

\item{type}{type of calculation ("T_MRCA" and later S_total)}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function and random generation for
the continuous phase-type distribution
}
\section{Functions}{
\itemize{
\item \code{dphtype}: Density function.

\item \code{qphtype}: Quantile Function.

\item \code{pphtype}: Distribution function.

\item \code{rphtype}: Random number generator.
So far only works for T_MRCA
}}

