% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phtype_functions.R
\name{dphtype}
\alias{dphtype}
\alias{qphtype}
\alias{pphtype}
\alias{rphtype}
\title{Density, distribution function, quantile function and random generation for
phase-type distributions}
\usage{
dphtype(x, obj)

qphtype(p, obj)

pphtype(q, obj)

rphtype(n_samples, phase_type, granularity = 0.01)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{granularity}{distance between numbers drawable}

\item{phase_type}{an object of class \code{phase_type}.}

\item{n_samples}{number of samples to draw.}

\item{n_OTU}{number of OTUs to consider}

\item{type}{type of calculation ("T_MRCA" and later S_total)}
}
\description{
Density, distribution function, quantile function and random generation for
phase-type distributions
}
\section{Functions}{
\itemize{
\item \code{dphtype}: Density function.

\item \code{qphtype}: Quantile Function.

\item \code{pphtype}: Distribution function.

\item \code{rphtype}: Random number generator.
}}

