% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itons.R
\name{itons}
\alias{itons}
\title{I-tons (frequency counts)}
\usage{
itons(mph_obj, i, theta)
}
\arguments{
\item{mph_obj}{Multivariate Phase Type object generated either from mult_phase_type or kingsman function}

\item{i}{Frequency count (positive integer)}

\item{theta}{mutation parameter (positive)}
}
\value{
A `disc_phase_type` object containing subintensity matrix (P), vector of initial probabilities (alpha) and defect (probability of not entering any transient
state prior to absorption)
}
\description{
Generate subintensity matrix, vector of initial probabilites and defect for any frequency count
}
\examples{
itons(kingsman(4), i = 2, theta = 2)

}
