% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kingsman.R
\name{kingsman}
\alias{kingsman}
\title{Simulate Matrices from Kingsman coalescent}
\usage{
kingsman(n)
}
\arguments{
\item{n}{sample size (a positive integer)}
}
\value{
A \code{mult_phase_type} object containing the subintensity matrix, reward matrix and vector of initial probabilities
}
\description{
Wrapper function that computes the matrices describing Kingsman coalescent ready for reward
transformation.
}
\examples{
kingsman(4)

}
