% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_type_constructor.R
\name{phase_type}
\alias{phase_type}
\title{Constructor function for the class \code{phase_type}.}
\usage{
phase_type(type = NULL, n = NULL, subint_mat = NULL,
  init_probs = NULL)
}
\arguments{
\item{type}{\code{'T_MRCA'}, \code{'T_Total'} or \code{NULL} (default).}

\item{n}{integer larger than 1 or \code{NULL} (default).}

\item{subint_mat}{matrix or \code{NULL} (default).}

\item{init_probs}{vector, a one-row matrix or \code{NULL} (default).}
}
\description{
Constructor function for the class \code{phase_type}.
}
