% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_phase_type_helper_functions.R
\name{RateMAndStateSpace}
\alias{RateMAndStateSpace}
\title{RateMAndStateSpace}
\usage{
RateMAndStateSpace(n)
}
\arguments{
\item{n}{integer larger than 1.}
}
\description{
Rate Matrix with corresponding state space
}
