% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_phase_type_helper_functions.R
\name{rewardtransformparm}
\alias{rewardtransformparm}
\title{Functions used to transform to discrete phase type and calculate the number of segregating sites
RewardTransformParm}
\usage{
rewardtransformparm(n, init_probs, subint_mat)
}
\description{
Function to compute reward transformed matrix T* as well as vector of initial probabilities
and the defect
}
