% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardtransformparm.R
\name{rewardtransformparm}
\alias{rewardtransformparm}
\title{Functions used to transform to discrete phase type and calculate the number of segregating sites
RewardTransformParm}
\usage{
rewardtransformparm(rewards, init_probs, subint_mat)
}
\arguments{
\item{rewards}{reward vector}

\item{init_probs}{initial probabilities}

\item{subint_mat}{sub-intensity matrix}
}
\description{
Function to compute reward transformed matrix T* as well as vector of initial probabilities
and the defect
}
