% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsites.R
\name{segsites}
\alias{segsites}
\title{Segregating sites}
\usage{
segsites(n, theta)
}
\arguments{
\item{n}{sample size (positive integer)}

\item{theta}{mutation parameter (positive)}
}
\value{
A `disc_phase_type` object containing subintensity matrix (P), vector of initial probabilities (alpha) and defect (probability of not entering any transient
state prior to absorption)
}
\description{
Generate subintensity matrix for the special case of segregating sites
}
\examples{
segsites(n = 4, theta = 2)

}
